+++ app/app/handlers/forms/moderator/add_view_job.py
@@ -6,6 +6,10 @@ from app.states.base import BaseStates
 from app.states.tgbot_states import AddViewWork
 from app.utils import const, get_data
 from app.utils.const import EDIT_WORK, EDIT_SORT, EDIT_SUBSISTEMS, FIO, ROLE, R_TYPE
+from dependency_injector.wiring import inject, Provide
+from app.services.application import ApplicationService
+from app.core.container import Container
+from app.models.application import Application
 
 
 async def get_sub_object(message: types.Message, state: FSMContext):
@@ -112,7 +116,9 @@ async def correct(query: types.CallbackQuery, state: FSMContext):
     await query.answer()
 
 
-async def edit(message: types.Message, state: FSMContext):
+@inject
+async def edit(message: types.Message, state: FSMContext,
+               application: ApplicationService = Provide[Container.application_service]):
     data = await state.get_data()
     point = data['change']
     if point == 'type_work':
@@ -123,47 +129,76 @@ async def edit(message: types.Message, state: FSMContext):
         await state.update_data(field_one=message.text)
     elif point == 'sort':
         await state.update_data(field_three=message.text)
-    new_kb = kb.sure().add(kb.exit_button)
-    await get_data.send_data(message=message, state=state)
-    await message.answer(const.SURE,
-                         reply_markup=new_kb)
-    await state.set_state(AddViewWork.sure)
+    data = await state.get_data()
+    if 'admin' not in data:
+        await get_data.send_data(message=message, state=state)
+        new_kb = kb.sure().add(kb.exit_button)
+        await message.answer(const.SURE, reply_markup=new_kb)
+        await state.set_state(AddViewWork.sure)
+    else:
+        if 'name' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'name': data['name']})
+        elif 'field_one' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'field_one': data['field_one']})
+        elif 'field_two' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'field_two': data['field_two']})
+        elif 'field_three' in data:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'field_three': data['field_three']})
+        await message.answer(const.CHANGE_SUCCESS)
+        await state.finish()
 
 
-async def get_subsystems_edit(query: types.CallbackQuery, state: FSMContext):
+@inject
+async def get_subsystems_edit(query: types.CallbackQuery, state: FSMContext,
+                              application: ApplicationService = Provide[Container.application_service]):
     new_kb = kb.accept().add(kb.exit_button)
     data = await state.get_data()
     if query.data != 'accept' and data['field_four'] == '':
         await state.update_data(field_four=query.data)
         data = await state.get_data()
-        msg = await query.message.answer(data['field_four'], reply_markup=new_kb)
+        msg = await query.message.answer(
+            data['field_four'], reply_markup=new_kb)
         await state.update_data(message_id=msg.message_id)
     elif query.data != 'accept':
         data = await state.get_data()
-        await state.update_data(field_four=data['field_four'] + ', ' + query.data)
+        await state.update_data(
+            field_four=data['field_four'] + ', ' + query.data)
         new_data = await state.get_data()
-        await bot.edit_message_text(new_data['field_four'],
-                                    query.message.chat.id,
-                                    new_data['message_id'], reply_markup=new_kb)
+        await bot.edit_message_text(
+            new_data['field_four'],
+            query.message.chat.id,
+            new_data['message_id'], reply_markup=new_kb)
     else:
-        await bot.delete_message(
-            query.message.chat.id, query.message.message_id)
-        await get_data.send_data(query=query, state=state)
-        new_kb = kb.sure().add(kb.exit_button)
-        await query.message.answer(const.SURE,
-                                   reply_markup=new_kb)
-        await state.set_state(AddViewWork.sure)
+        data = await state.get_data()
+        if 'admin' not in data:
+            await bot.delete_message(
+                query.message.chat.id, query.message.message_id)
+            await get_data.send_data(query=query, state=state)
+            new_kb = kb.sure().add(kb.exit_button)
+            await query.message.answer(const.SURE, reply_markup=new_kb)
+            await state.set_state(AddViewWork.sure)
+        else:
+            await application.update(data['admin'],
+                                     obj_in={'application_status': Application.ApplicationStatus.in_work,
+                                             'field_four': data['field_four']})
+            await query.message.answer(const.CHANGE_SUCCESS)
+            await state.finish()
     await query.answer()
 
 
-async def get_role(query: types.CallbackQuery, state: FSMContext):
+@inject
+async def get_role(query: types.CallbackQuery, state: FSMContext,
+                   application: ApplicationService = Provide[Container.application_service]):
     await bot.delete_message(query.message.chat.id, query.message.message_id)
     await state.update_data(role=query.data)
-    new_kb = kb.sure().add(kb.exit_button)
-    await get_data.send_data(query=query, state=state)
-    await query.message.answer(const.SURE,
-                               reply_markup=new_kb)
-    await state.set_state(AddViewWork.sure)
 
 
 def register(dp: Dispatcher):
